cursor_mcp/weather/requirements.txt
rich>=10.0.0
typer>=0.9.0
pydantic>=2.0.0
pyyaml>=6.0.0 